+++ sdk/servicebus/service-bus/src/serializers/ruleResourceSerializer.ts
@@ -44,14 +44,14 @@ function getTopicFilter(value: any): SqlRuleFilter | CorrelationRuleFilter {
   } else {
     result = {
       correlationId: getStringOrUndefined(value["CorrelationId"]),
-      label: getStringOrUndefined(value["Label"]),
+      subject: getStringOrUndefined(value["Label"]),
       to: getStringOrUndefined(value["To"]),
       replyTo: getStringOrUndefined(value["ReplyTo"]),
       replyToSessionId: getStringOrUndefined(value["ReplyToSessionId"]),
       sessionId: getStringOrUndefined(value["SessionId"]),
       messageId: getStringOrUndefined(value["MessageId"]),
       contentType: getStringOrUndefined(value["ContentType"]),
-      properties: getKeyValuePairsOrUndefined(value["Properties"], "UserProperties")
+      applicationProperties: getKeyValuePairsOrUndefined(value["Properties"], "UserProperties")
     };
   }
   return result;
@@ -189,14 +189,17 @@ export class RuleResourceSerializer implements AtomXmlSerializer {
 
         resource.Filter = {
           CorrelationId: correlationFilter.correlationId,
-          Label: correlationFilter.label,
+          Label: correlationFilter.subject,
           To: correlationFilter.to,
           ReplyTo: correlationFilter.replyTo,
           ReplyToSessionId: correlationFilter.replyToSessionId,
           ContentType: correlationFilter.contentType,
           SessionId: correlationFilter.sessionId,
           MessageId: correlationFilter.messageId,
-          Properties: buildInternalRawKeyValuePairs(correlationFilter.properties, "userProperties")
+          Properties: buildInternalRawKeyValuePairs(
+            correlationFilter.applicationProperties,
+            "userProperties"
+          )
         };
         resource.Filter[Constants.XML_METADATA_MARKER] = {
           "p4:type": "CorrelationFilter",