+++ sdk/servicebus/service-bus/test/renewLock.spec.ts
@@ -164,7 +164,7 @@ describe("Message Lock Renewal", () => {
 
     await delay(5000);
     if (msgs[0].lockToken) {
-      await msgs[0].renewMessageLock();
+      await msgs[0].renewLock();
     }
 
     // Compute expected lock expiry time after renewing lock after 5 seconds
@@ -262,7 +262,7 @@ describe("Message Lock Renewal", () => {
         );
 
         await delay(5000);
-        await brokeredMessage.renewMessageLock();
+        await brokeredMessage.renewLock();
 
         // Compute expected lock expiry time after renewing lock after 5 seconds
         expectedLockExpiryTimeUtc.setSeconds(expectedLockExpiryTimeUtc.getSeconds() + 5);